+++ package.json
@@ -17,14 +17,17 @@
     "@tailwindcss/forms": "^0.5.10",
     "@tailwindcss/line-clamp": "^0.4.4",
     "@tailwindcss/typography": "^0.5.16",
+    "express": "^4.21.2",
     "mapbox-gl": "^3.9.4",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-loader-spinner": "^6.1.6",
     "react-loading-indicators": "^1.0.0",
     "react-map-gl": "^7.1.9",
+    "socket.io": "^4.8.1",
     "socket.io-client": "^4.8.1",
     "systeminformation": "^5.25.11",
+    "ws": "^8.18.0",
     "zustand": "^5.0.3"
   },
   "devDependencies": {
