+++ src/Pages/Home.jsx
@@ -7,6 +7,7 @@ import FlightControlPannel from "../Components/FlightControlPannel/FlightControl
 import Map from "../Components/Map/Map";
 import StatusBox from "../Components/Statusbox/Statusbox";
 // import InputParameter from "../Components/InputParameter/InputParameter";
+import Text from "../api/text";
 function MainPage() {
   const [count, setCount] = useState(0);
 
@@ -42,6 +43,12 @@ function MainPage() {
           className="w-screen h-[30px] border-2 relative bottom-0"
         ></div>
       </div>
+
+
+
+
+      <Text />
+
     </>
   );
 }
