+++ src/api/config.js
@@ -1,1 +1,30 @@
-const backedApi= 'http://localhost:3000/api';
+const express = require("express");
+const http = require("http");
+const { Server } = require("socket.io");
+
+const app = express();
+const server = http.createServer(app);
+
+const io = new Server(server, {
+  cors: {
+    origin: "http://localhost:5173", // Allow requests from your React app
+    methods: ["GET", "POST"],
+  },
+});
+
+io.on("connection", (socket) => {
+  console.log("A user connected:", socket.id);
+
+  socket.on("message", (data) => {
+    console.log("Message from client:", data);
+    socket.emit("response", `Server received: ${data}`);
+  });
+
+  socket.on("disconnect", () => {
+    console.log("User disconnected");
+  });
+});
+
+server.listen(5000, () => {
+  console.log("Server running on http://localhost:5000");
+});
