+++ src/api/text.jsx
@@ -1,0 +1,52 @@
+
+import React, { useEffect, useState } from 'react';
+import io from 'socket.io-client';
+
+const socket = io('http://localhost:5000'); // Adjust the URL as needed
+
+const ChatComponent = () => {
+    const [messages, setMessages] = useState([]);
+    const [input, setInput] = useState('');
+
+    useEffect(() => {
+        socket.on('connect', () => {
+            console.log('Connected to server');
+        });
+
+        socket.on('message', (message) => {
+            setMessages((prevMessages) => {
+
+            [...prevMessages, message]
+            console.log('Message received:', prevMessages);
+            }
+        );
+        });
+
+        return () => {
+            socket.disconnect();
+        };
+    }, []);
+
+    const sendMessage = () => {
+        socket.send(input);
+        setInput('');
+    };
+
+    return (
+        <div className='border-2 flex absolute top-48 left-7'>   
+            <div>
+                {messages.map((msg, index) => (
+                    <div key={index}>{msg}</div>
+                ))}
+            </div>
+            <input
+                value={input}
+                onChange={(e) => setInput(e.target.value)}
+                placeholder="Type a message..."
+            />
+            <button onClick={sendMessage}>Send</button>
+        </div>
+    );
+};
+
+export default ChatComponent;
